remove task

+++ src/app/pages/todo-list/components/to-do-list/to-do-list.component.ts
@@ -1,13 +1,16 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, HostListener, OnDestroy, OnInit} from '@angular/core';
 import {AngularFirestore} from '@angular/fire/compat/firestore';
 import {TaskService} from '../../services/task.service';
-import {combineLatest, Observable, of} from 'rxjs';
+import {combineLatest, Observable, of, Subscription} from 'rxjs';
 import {Task} from '../../models/task-model';
 import {AngularFireAuth} from '@angular/fire/compat/auth';
 import {CreateItemComponent} from '../create-item/create-item.component';
 import {MatDialog} from '@angular/material/dialog';
 import {CdkDragDrop, moveItemInArray} from '@angular/cdk/drag-drop';
 import {AuthService} from '../../../auth/services/auth.service';
+import {DialogConfirmComponent} from '../../../../shared/components/dialog-confirm/dialog-confirm.component';
+import {MatBottomSheet} from '@angular/material/bottom-sheet';
+import {ConfirmBottomSheetComponent} from '../../../../shared/components/confirm-bottom-sheet/confirm-bottom-sheet.component';
 
 
 @Component({
@@ -15,13 +18,20 @@ import {AuthService} from '../../../auth/services/auth.service';
   templateUrl: './to-do-list.component.html',
   styleUrls: ['./to-do-list.component.scss']
 })
-export class ToDoListComponent implements OnInit {
+export class ToDoListComponent implements OnInit, OnDestroy {
   tasks: Task[];
+  subscriptions: Subscription = new Subscription();
+  screenMode;
+  public getScreenWidth: any;
+  public getScreenHeight: any;
+
 
   constructor(private db: AngularFirestore, private taskService: TaskService,
               public authService: AuthService,
               public afAuth: AngularFireAuth,
-              public dialog: MatDialog) {
+              public dialog: MatDialog,
+              private bottomSheet: MatBottomSheet
+  ) {
     this.tasks = [];
   }
 
@@ -30,7 +40,18 @@ export class ToDoListComponent implements OnInit {
   // }
 
   ngOnInit(): void {
+    this.getScreenWidth = window.innerWidth;
+    this.getScreenHeight = window.innerHeight;
+    this.subscriptions.add(this.taskService.taskFiltered$.asObservable().subscribe(tasks => {
+      this.tasks = tasks;
+    }));
     this.getTasks();
+  }
+
+  @HostListener('window:resize', ['$event'])
+  onWindowResize() {
+    this.getScreenWidth = window.innerWidth;
+    this.getScreenHeight = window.innerHeight;
   }
 
   getTasks() {
@@ -73,7 +94,27 @@ export class ToDoListComponent implements OnInit {
   }
 
   onRemove(task) {
-    this.taskService.removeTask(task.id).finally();
+    if (this.getScreenWidth > 959) {
+      const dialog = this.dialog.open(DialogConfirmComponent, {
+        disableClose: true,
+      });
+      dialog.afterClosed().subscribe((success) => {
+        if (success) {
+          this.taskService.removeTask(task.id).finally();
+
+        }
+      });
+    } else {
+      const bSheet = this.bottomSheet.open(ConfirmBottomSheetComponent, {
+        disableClose: true,
+      });
+      bSheet.afterDismissed().subscribe((task) => {
+      });
+    }
+  }
+
+  ngOnDestroy() {
+    this.subscriptions.unsubscribe();
   }
 
 }
